Fix osx-arm64 compilation
Some files, like COS libraries, were missing in the osx-arm64 package.

Add patch to force appl and services compilation.
In 4.2.x, cross-compiling is activated with make variables calling it
an "EmbeddedSystem", with the expectation that it is a minimal small
system, and therefore it does not build things like omniNames or the
COS libraries.

See https://www.omniorb-support.com/pipermail/omniorb-list/2022-October/032229.html