Add patch to find openssl root
openssl 3.3.0 doesn't include prefix in the pkgconfig file.

pkg-config finds openssl but "pkg-config --variable=prefix openssl" returns an empty string.
OPEN_SSL_ROOT is empty and it prevents ssl transport to be compiled (even if openssl is reported as found).

Patch used to return the prefix when that value is empty.